requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
langchain-openai==0.3.0
langgraph==0.2.60
pydantic==2.10.3
scikit-learn==1.6.0
numpy==2.2.1
python-dotenv==1.0.1
flask==3.1.0
